FIX: removed compiler warnings (for VC 7.1, at least)
git-svn-id: https://svn.code.sf.net/p/flashfolder/code/trunk@63 a429ba1a-122f-0410-b780-ea7094eed80e

+++ common/HistoryLst.h
@@ -33,8 +33,8 @@ public:
 	void AddFolder( LPCTSTR s );
 	
     void SetMaxEntries( int count );
-    inline int GetMaxEntries() const { return m_maxEntries; };
-    inline int GetCount() const { return m_list.size(); };
+    inline size_t GetMaxEntries() const { return m_maxEntries; };
+    inline size_t GetCount() const { return m_list.size(); };
 	
     inline const tstring &operator [](int n) const { return m_list[n]; };
     inline const std::vector<tstring> &GetList() const { return m_list; };
